feat: booking pages read bundle config from DB via server components
Server components fetch active bundleConfig and pass it as props; BundleForm extracted as new client component so bundle page can redirect when no config is active.
